Renamed plot_jupyter to plot_plotly Automated activation of logging when module is loaded
