+++ __init__.py
@@ -1,0 +1,4 @@
+import util
+
+name = "PySME"
+util.start_logging()
